Added bracket test
